add ability to use different state management strategies (SMS)
To create websites / microservices with no shared session, OAuth state
must be maintained differently. Besides the prior session based SMS,
this commit adds a signed token SMS. This will allow a login service
to authenticate users based upon OAuth and hand the users to a
different service (after redirect) or to a different instance of the
same service. All you need to share is the public key of the login
service on startup.